moved all MatchingStatus/Listener functionality under separate module
